[Relation Viewer] update logic to rerender viewer when node click
this commit creates various functions such as displaying deleted file, add a icon instead of text, change theme etc...

ONE-vscode-DCO-1.0-Signed-off-by: jihong-yu <jimdac@naver.com>
