+++ media/MetadataViewer/style.css
@@ -26,32 +26,20 @@ body.vscode-light {
 }
 
 body.vscode-light .view-item-header{
-  color: #252526;
+  color: #444444;
 }
 
 body.vscode-light .view-item-name {
-  color: #242424;
+  color: #616161;
 }
 
 body.vscode-light .view-item-sub-header{
-  color: steelblue;
+  color: #a31515;
 }
 
 body.vscode-light .view-item-value{
-  color: black;
+  color: #444444;
   background-color: white;
-}
-
-body.vscode-light .breadcrumbs-relativepath {
-  background-color: #ececec;
-}
-
-body.vscode-light .breadcrumbs-coverbar {
-  background-color: #ececec;
-}
-
-body.vscode-light .scrolled {
-  box-shadow: 0px 3px 10px #888888F0;
 }
 
 .view-item-header {
@@ -61,10 +49,8 @@ body.vscode-light .scrolled {
   font-size: 20px;
   font-weight: bold;
   font-family: sans-serif;
-  color: lightblue;
+  color: #e7e7e7;
 }
-
-
 
 .view-item-content {
   display: flex;
@@ -89,18 +75,18 @@ body.vscode-light .scrolled {
 }
 
 .view-item-name {
-  color: navajowhite;
+  color: #cccccc;
   min-width: 100px;
 }
 
 .view-item-content{
-  color:white;
+  color:#e7e7e7;
 }
 
 .view-item-sub-header{
   margin-left: 10px;
   font-size: 16px;
-  color: white;
+  color: #d7ba7d;
   margin-bottom: 10px;
   font-weight: bold;
 }
