+++ package.json
@@ -217,12 +217,12 @@
         "category": "ONE"
       },
       {
-        "command": "one.metadata.showMetadataViewer",
+        "command": "one.viewer.metadata.show",
         "title": "Show Metadata",
         "category": "ONE"
       },
       {
-        "command": "one.relation.showRelationViewer",
+        "command": "one.viewer.relation.show",
         "title": "Show Relation",
         "category": "ONE"
       },
@@ -312,25 +312,25 @@
           "group": "inline"
         },
         {
-          "command": "one.metadata.showMetadataViewer",
+          "command": "one.viewer.metadata.show",
           "when": "view == OneExplorerView && viewItem =~ /baseModel|product/",
           "group": "7_metadata@1"
         },
         {
-          "command": "one.relation.showRelationViewer",
+          "command": "one.viewer.relation.show",
           "when": "view == OneExplorerView && viewItem =~ /baseModel|product/",
           "group": "7_metadata@2"
         }
       ],
       "explorer/context": [
         {
-          "command": "one.metadata.showMetadataViewer",
-          "when": "resourceExtname in metadata.supportedFiles && !explorerResourceIsFolder",
+          "command": "one.viewer.metadata.show",
+          "when": "resourceExtname in one.metadata.supportedFiles && !explorerResourceIsFolder",
           "group": "7_metadata@1"
         },
         {
-          "command": "one.relation.showRelationViewer",
-          "when": "resourceExtname in relation.supportedFiles && !explorerResourceIsFolder",
+          "command": "one.viewer.relation.show",
+          "when": "resourceExtname in one.relation.supportedFiles && !explorerResourceIsFolder",
           "group": "7_metadata@2"
         }
       ],
