+++ src/RelationViewer/RelationViewerProvider.ts
@@ -17,7 +17,7 @@
 import * as vscode from 'vscode';
 import { Node } from '../OneExplorer/OneExplorer';
 import { RelationViewer } from './RelationViewer';
-
+import {getRelationData} from './example/RelationExample';
 
 export class RelationViewerDocument implements vscode.CustomDocument {
   private readonly _uri: vscode.Uri;
@@ -53,12 +53,9 @@ export class RelationViewerDocument implements vscode.CustomDocument {
   public openView(panel: vscode.WebviewPanel, extensionUri: vscode.Uri, fileUri:vscode.Uri) {
     let view = new RelationViewer(panel,extensionUri);
 
-    view.initRelationViewer();
+    view.initWebview();
     view.loadContent();
     this._metadataViwer.push(view);
-
-    //get Relative_Path
-    //const relativePath:string = vscode.workspace.asRelativePath(fileUri);
     
     const payload = getRelationData(fileUri);
 
@@ -96,7 +93,7 @@ export class RelationViewerProvider implements
           retainContextWhenHidden: true,
         }
       }),
-      vscode.commands.registerCommand('one.relation.showRelationViewer', async (uri) => {
+      vscode.commands.registerCommand('one.viewer.relation.show', async (uri) => {
         //If the method is executed in the ONE explorer, change the uri.
         let fileUri = uri;
         if(uri instanceof Node){
@@ -109,7 +106,7 @@ export class RelationViewerProvider implements
     ];
     
     // supported file extension to show relations context menu
-    vscode.commands.executeCommand('setContext', 'relation.supportedFiles', [
+    vscode.commands.executeCommand('setContext', 'one.relation.supportedFiles', [
       '.tflite',
       '.pb',
       '.onnx',
@@ -144,49 +141,4 @@ export class RelationViewerProvider implements
       _token: vscode.CancellationToken): Promise<void> {
     document.openView(webviewPanel, this._context.extensionUri, document.uri);
   }
-}
-
-export function getRelationData(path:any) {
-  
-  const dummyData = {
-    "selected": "1",
-    "relation-data": [
-      {"id": "1", "parent": "", "represent-idx": 0, "data-list": [{"name": "baseModelTestTflite123123.tflite", "path": "baseModelTestTflite123123.tflite", "is-deleted":false},{"name": "model.tflite", "path": "model.tflite", "is-deleted":true},{"name": "c.tflite", "path": "c.tflite"},{"name": "d.tflite", "path": "d.tflite", "is-deleted":false}]},  // TODO: id, parentid: hashId
-      {"id": "2", "parent": "1", "represent-idx": 0, "data-list": [{"name": "test1.circle", "path": "src/hello/test1.circle", "onecc-version": "1.0.0", "toolchain-version": "1.0.0", "is-deleted":false}]},
-      {"id": "3", "parent": "2", "represent-idx": 0, "data-list": [{"name": "test2.circle", "path": "src/trudiv/model/test2.circle", "onecc-version": "1.0.0", "toolchain-version": "1.0.0", "is-deleted":false}]},
-      {"id": "4", "parent": "1", "represent-idx": 0, "data-list": [{"name": "test1.log", "path": "test1.log", "onecc-version": "1.0.0", "toolchain-version": "1.0.0", "is-deleted":false}]},
-      {"id": "5", "parent": "2", "represent-idx": 0, "data-list": [{"name": "test2.log", "path": "test2.log", "toolchain-version": "1.0.0", "is-deleted":false}]},
-      {"id": "6", "parent": "4", "represent-idx": 0, "data-list": [{"name": "baseModelTestCircle.circle", "path": "baseModelTestCircle.circle", "is-deleted":false}]},
-      {"id": "7", "parent": "6", "represent-idx": 0, "data-list": [{"name": "model.q8.circle", "path": "model.q8.circle", "onecc-version": "1.0.0", "toolchain-version": "1.0.0", "is-deleted":false}]},
-      {"id": "8", "parent": "6", "represent-idx": 0, "data-list": [{"name": "pbTestCircle1.log", "path": "pbTestCircle1.log", "onecc-version": "1.0.0", "toolchain-version": "1.0.0", "is-deleted":false}]},
-      {"id": "9", "parent": "7", "represent-idx": 0, "data-list": [{"name": "test_onnx.circle", "path": "hello/test_onnx.circle", "toolchain-version": "1.0.0", "is-deleted":false}]},
-      {"id": "10", "parent": "7", "represent-idx": 0, "data-list": [{"name": "while_000.circle", "path": "while/while_000.circle", "onecc-version": "1.0.0", "toolchain-version": "1.0.0", "is-deleted":false}]},
-      {"id": "11", "parent": "8", "represent-idx": 0, "data-list": [{"name": "e1.log", "path": "e1.circle", "onecc-version": "1.0.0", "toolchain-version": "1.0.0", "is-deleted":false}]},
-      {"id": "12", "parent": "8", "represent-idx": 0, "data-list": [{"name": "e2.log", "path": "e2.circle", "onecc-version": "1.0.0", "toolchain-version": "1.0.0", "is-deleted":true},{"name": "e3.circle", "path": "e3.circle", "onecc-version": "1.2.0", "toolchain-version": "1.0.0", "is-deleted":false}]}
-    ]
-  } as any;
-
-  for (const key in dummyData) {
-    if(key === 'relation-data'){
-      for (const idx in dummyData['relation-data']) {
-        for (const key2 in dummyData['relation-data'][idx]) {
-            if(key2 === 'data-list'){
-              for (let index = 0; index < dummyData['relation-data'][idx]['data-list'].length; index++) {
-                const element = dummyData['relation-data'][idx]['data-list'][index];
-                for (const key3 in element) {
-                  if(key3 === 'path'){
-                    if(element['path'] === path){
-                      dummyData['relation-data'][idx]['represent-idx'] = index;
-                      dummyData['selected'] = dummyData['relation-data'][idx]['id'];
-                    }
-                  }
-                }
-              }
-            }
-        }
-      }
-    }
-  }
-
-  return dummyData;
 }
